+++ game/src/dlls/objects/206/206.c
@@ -669,7 +669,7 @@ int dll_CE_getObjectTypeId(void) {
 void dll_CE_free(GameObject* obj) {
     GroundBaddieState* state = obj->extra;
 
-    objFreeObjectType((int)obj, DLL_CE_OBJGROUP);
+    objFreeObjectType(obj, DLL_CE_OBJGROUP);
     {
         GameObject* child = obj->childObjs[0];
