+++ game/src/dlls/objects/211/211.c
@@ -1089,7 +1089,7 @@ int dll_D3_getObjectTypeId(void)
 void dll_D3_free(GameObject* obj)
 {
     int* inner = obj->extra;
-    objFreeObjectType((int)obj, STAFFACTION_OBJGROUP);
+    objFreeObjectType(obj, STAFFACTION_OBJGROUP);
     if (obj->childObjs[0] != NULL)
     {
         Obj_FreeObject(obj->childObjs[0]);
