+++ game/src/dlls/objects/231/231.c
@@ -68,7 +68,7 @@ int FlammableVine_getObjectTypeId(void) {
 }
 
 void FlammableVine_free(GameObject* obj) {
-    objFreeObjectType((int)obj, FLAMMABLEVINE_OBJECT_GROUP);
+    objFreeObjectType(obj, FLAMMABLEVINE_OBJECT_GROUP);
 }
 
 void FlammableVine_render(GameObject* obj, int fwdArg2, int fwdArg3, int fwdArg4, int fwdArg5, s8 visible) {
@@ -199,7 +199,7 @@ void FlammableVine_init(GameObject* obj, FlammableVinePlacement* placement) {
     f32 scale;
 
     state = obj->extra;
-    objAddObjectType((int)obj, FLAMMABLEVINE_OBJECT_GROUP);
+    objAddObjectType(obj, FLAMMABLEVINE_OBJECT_GROUP);
     obj->anim.rotX = (s16)(placement->rotXByte << 8);
 
     obj->anim.rootMotionScale =
