+++ game/src/dlls/objects/254_MagicPlant/MagicPlant.c
@@ -251,8 +251,8 @@ void MagicPlant_free(GameObject* obj, int keepChildren) {
     MagicPlantState* state;
 
     state = obj->extra;
-    objFreeObjectType((int)obj, MAGICPLANT_OBJGROUP_A);
-    objFreeObjectType((int)obj, MAGICPLANT_OBJGROUP_B);
+    objFreeObjectType(obj, MAGICPLANT_OBJGROUP_A);
+    objFreeObjectType(obj, MAGICPLANT_OBJGROUP_B);
     if (obj->childCount != 0) {
         ObjLink_DetachChild(obj, state->childObject);
         if (keepChildren == 0) {
@@ -391,8 +391,8 @@ void MagicPlant_init(GameObject* obj, MagicPlantPlacement* placement) {
 
     anim = &obj->anim;
     state = obj->extra;
-    objAddObjectType((int)obj, MAGICPLANT_OBJGROUP_A);
-    objAddObjectType((int)obj, MAGICPLANT_OBJGROUP_B);
+    objAddObjectType(obj, MAGICPLANT_OBJGROUP_A);
+    objAddObjectType(obj, MAGICPLANT_OBJGROUP_B);
     noSaveTime = (*gMapEventInterface)->shouldNotSaveTime(placement->eventId);
     if (noSaveTime == 0) {
         progress = (*gMapEventInterface)->getTime(placement->eventId);
