+++ game/src/dlls/objects/454_DIMCannon/DIMCannon.c
@@ -431,7 +431,7 @@ void DIMCannon_free(GameObject* obj) {
         Resource_Release(gDimCannonResource);
         gDimCannonResource = NULL;
     }
-    objFreeObjectType((int)obj, DIM_CANNON_OBJECT_GROUP);
+    objFreeObjectType(obj, DIM_CANNON_OBJECT_GROUP);
 }
 
 void DIMCannon_render(GameObject* obj, int renderArg2, int renderArg3, int renderArg4, int renderArg5,
