+++ game/src/dlls/objects/472_DIM2PathGen/DIM2PathGen.c
@@ -115,7 +115,7 @@ void DIM2PathGenerator_update(GameObject* obj) {
             childPlacementData->base.ident = placement->base.ident;
             DIM2_SNOW_BALL_INTERFACE(objects[objectIndex])
                 ->init((GameObject*)objects[objectIndex], (Dim2SnowBallPlacement*)childPlacementData, 1);
-            objFreeObjectType((int)objects[objectIndex], DIM2_PATH_GENERATOR_SNOWBALL_GROUP);
+            objFreeObjectType(objects[objectIndex], DIM2_PATH_GENERATOR_SNOWBALL_GROUP);
             objGetAllOfType(DIM2_PATH_GENERATOR_SNOWBALL_GROUP, &count);
             for (poolIndex = 0; poolIndex < count; poolIndex++) {
             }
