+++ game/src/dlls/objects/480_DIM_Boss/DIM_Boss.c
@@ -1614,7 +1614,7 @@ void DIMboss_free(GameObject* obj) {
     mainSetBits(GAMEBIT_DIM_TriggerLostInBlizzard, 0);
     obj->anim.resetHitboxFlags &= ~DIMBOSS_OBJECT_FLAG_ACTIVE;
     CameraShake_Disable();
-    objFreeObjectType((int)obj, DIMBOSS_OBJGROUP);
+    objFreeObjectType(obj, DIMBOSS_OBJGROUP);
     childObject = obj->childObjs[0];
     if (childObject != NULL) {
         Obj_FreeObject(childObject);
